Modified assignments of snid in `CatSimSampling`:
Previously, the code was designed so that multiple supernovae could
be hosted by the same galaxy. We are changeing that to have allow only a
single supernova per galaxy. In these cases the snid can be equal to the
galtileid. Further, because we want to use the case where we are only
working on a single tile, it is sufficient to use id for this value
	modified:   snsims/paramDistribution.py
	modified:   snsims/version.py