Fixed issue with mjdmin = 59580
In __VERSION__ 0.7.0  and second release, the class `TwinklesSim`
supplied a hard coded value of 59580 to an instantiation of `SimpleSALTDist`
as the parameter `mjdmin`. This arose from the desire to distribute the
times of explosions `t0` during the time of the survey. If this is used
to generate SN parameters for the catsim database, it is actually useful
to set this to 0.0 The reason is that the catsim SN instance catalog has
a variable to do this offset, and can do it for different realizations
of OpSim which have different starting times. Also, having a hard coded
parameter is bad, so changing that to a parameter in the constructor of
`TwinkleSim` with a default value of 0.
	modified:   paramDistribution.py